Refactor user hook into separate file
